+++ packages/sdk/src/constants/CicsCmci.constants.ts
@@ -257,4 +257,8 @@ export const CicsCmciConstants = {
      */
     INVALIDDATA: 1041,
   },
+  /*
+  * Default Resp Code 
+  */
+  DEFAULT_RESP_CODE: "0000",
 };